Refactor into helpers package
