+++ packages/helpers/index.ts
@@ -27,6 +27,8 @@ export { default as isPlainObject } from './is-plain-object'
 export { default as isPerceivedBy } from './is-perceived-by'
 export { default as isValidUrl } from './is-valid-url'
 export { default as padInteger } from './padInteger'
+export { matchedForceAndRoleFilter, matchedAllRolesFilter, matchedForceFilter } from './participant-filters'
+export { checkParticipantStates, getParticipantStates } from './participant-states'
 export { default as plannedRouteFor } from './planned-route-for'
 export { default as roundToNearest } from './roundToNearest'
 export { default as routeAddSteps } from './route-add-steps'
