+++ packages/helpers/participant-states.ts
@@ -1,5 +1,5 @@
 import { PlayerUi, ChannelData, Participant } from '@serge/custom-types';
-import { matchedForceAndRoleFilter, matchedAllRolesFilter, matchedForceFilter } from '../helpers/filters';
+import { matchedForceAndRoleFilter, matchedAllRolesFilter, matchedForceFilter } from './participant-filters';
 
 export interface CheckParticipantStates {
   /** whether role is participant in channel */
